notification-service/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pymongo==4.16.0
python-dotenv==1.2.2
pydantic==2.12.5
